requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
analytics-python==1.4.post1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
audioread==3.0.1
backoff==1.10.0
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
fastapi==0.115.12
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
google-auth==2.39.0
googleapis-common-protos==1.70.0
gradio==3.0
gradio_client==1.10.0
greenlet==3.2.1
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
hnswlib==0.8.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.2
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.1.14
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.1
langchain-text-splitters==0.0.2
langsmith==0.1.147
lazy_loader==0.4
librosa==0.10.1
linkify-it-py==2.0.3
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.76.2
opencv-python==4.11.0.86
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.18
overrides==7.7.0
packaging==23.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
posthog==4.0.1
propcache==0.3.1
protobuf==5.29.4
pulsar-client==3.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.2.2
tqdm==4.67.1
transformers==4.40.0
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.4.0
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
yt-dlp==2025.4.30
zipp==3.21.0
